Support crc32 even if on arm32
CRC32 is supported on A32 and T32.
